refactor: axios client and query client
