+++ frontend/src/features/documents/api/documentApi.ts
@@ -1,4 +1,4 @@
-import api from "@/app/api";
+import api from "@/app/axiosClient";
 import {
   DocumentCategoryResponseSchema,
   DocumentTypeResponseSchema,
